Send notifications in background
The aim of this change is to make notifications to be sent in
background. In order to achieve that, a supervised GenServer was added.

Every time there's an exception, `BoomNotifier` will delegate the
sending process to the `NotifierSenderServer` making async
`GenServer.cast` calls.

This server will be responsible to spawn a new process that will execute
the `Notifier.notify` functions with the provided params.  It will also
log the errors if a notifier crashes.

This server is supervised so it gets restarted in case something fails.