Send notifications in background (#40)
* Send notifications in background

The aim of this change is to make notifications to be sent in
background. In order to achieve that, a supervised GenServer was added.

Every time there's an exception, `BoomNotifier` will delegate the
sending process to the `NotifierSenderServer` making async
`GenServer.cast` calls.

This server will be responsible to spawn a new process that will execute
the `Notifier.notify` functions with the provided params.  It will also
log the errors if a notifier crashes.

This server is supervised so it gets restarted in case something fails.

* Remove warning for unused var

* Wait for request to be made

* Increase receive timout for async asserts

* Change module assertion

* Test against latest possible versions of library deps (#42)

* Test against latest possible versions of library deps

* Ask GitHub to show result of all matrix jobs despite any failure

* Fix Phoenix warning after phoenix dep update

* No longer test against Erlang 20.0 unsupported by latest cowboy

* Fix race condition when waiting for the log to be catpured

* Refactor GenServer name

* Upgrade more deps (#45)

* Move to bamboo 2.x from 1.x

* Move to dialyxir 1.x from 0.x

* Move to bypass 2.x from 1.x

* Fix warning 'warning: variable name is unused' (#43)

* Change supervisor strategy

Co-authored-by: Gonzalo <grzuy0@gmail.com>