Provide callback when ErrorHandler is already in use
Boom was always relying on ErrorHandler plug, but it was
conflicting when the host application is already using that plug.

Now, we are checking at compilation time if the application Router
module already has an implementation for handle_errors/2 (because it is
provided when using ErrorHandler module, if not implemented by the
host router module).

When handle_errors/2 implementation is controlled by the application,
library's users are responsible to call notify_error/2 so
notifications are effectively produced.