Refactor `ErrorStorage` module
Currently, we are grouping every error information in several lists
according to its error type. Eventually, this could consume a lot of
memory without providing any important benefit (the only piece of
information that changes is the timestamp and the request data).

This commit replaces that with a map structure where its key is the
hashed error info and its value is the actual error info plus:
  * the number of times this error occurred,
  * the first and the list time it happened.